% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treeModule.R
\name{is.uniq}
\alias{is.uniq}
\title{Function to filter unique variables}
\usage{
is.uniq(data)
}
\arguments{
\item{i}{variable}
}
\value{
not unique column names
}
\description{
to prevent error when regression's y value is unique
}
