% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_groupStatModule.R
\name{mod_groupStatModule_server}
\alias{mod_groupStatModule_server}
\title{Shiny Module for data transformation}
\usage{
mod_groupStatModule_server(id, inputData)
}
\arguments{
\item{id}{id of module}

\item{inputData}{"reactive" data}
}
\value{
transformed data
}
\description{
groupStatModule Server Functions
}
