% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_groupStatModule.R
\name{mod_groupStatModule_ui}
\alias{mod_groupStatModule_ui}
\title{Shiny Module for data transformation}
\usage{
mod_groupStatModule_ui(id)
}
\arguments{
\item{id}{id of module}
}
\description{
groupStatModule UI Function
}
