% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_kmsModule.R
\name{mod_kmsModule_server}
\alias{mod_kmsModule_server}
\title{Shiny Module for k means clustering}
\usage{
mod_kmsModule_server(id, inputData)
}
\arguments{
\item{id}{id of module}

\item{inputData}{"reactive" data}
}
\description{
kmsModule Server Functions
}
