% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pcaModule.R
\name{mod_pcaModule_server}
\alias{mod_pcaModule_server}
\title{Shiny Module for data transformation}
\usage{
mod_pcaModule_server(id, inputData)
}
\arguments{
\item{id}{id of module}

\item{inputData}{"reactive" data}
}
\description{
pcaModule Server Functions
}
