% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treeModule.R
\name{mod_treeModule_server}
\alias{mod_treeModule_server}
\title{Shiny Module for data transformation}
\usage{
mod_treeModule_server(id, inputData)
}
\arguments{
\item{id}{id of module}

\item{inputData}{"reactive" data}
}
\value{
transformed data
}
\description{
treeModule Server Functions
}
