% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treeModule.R
\name{mod_treeModule_ui}
\alias{mod_treeModule_ui}
\title{Shiny Module for data transformation}
\usage{
mod_treeModule_ui(id)
}
\arguments{
\item{id}{id of module}
}
\description{
treeModule UI Function
}
